+++ resources/cli/templates/install/service/veil/resources/views/examples/layouts/partials/footer.veil.php
@@ -16,6 +16,4 @@
  *
  */
 ?>
-<footer id="footer" style="background-color:#3b82f6;padding:1rem;color:#fff;text-align:center;font-size:0.9rem;">
-    &copy; {{year}} Website Name &middot; All rights reserved
-</footer>
+<footer>Copyright &copy; {{year}} &middot; All rights reserved.</footer>