+++ src/Application/Kernel/Console/Commands/InstallService.php
@@ -67,7 +67,7 @@ class InstallService extends Command
 
                 // Add PDO extension to Composer
 
-                shell_exec('composer require ext-pdo');
+                //shell_exec('composer require ext-pdo');
 
                 ConsoleUtilities::msgInstalled($name, $output);
 
@@ -125,7 +125,8 @@ class InstallService extends Command
 
             ConsoleUtilities::msgInstallComplete('Database service', $output);
 
-            $output->writeln('<info>*** Be sure to update .env with your database credentials! ***</info>');
+            $output->writeln('<info>*** NOTE: Be sure to update .env with your database credentials! ***</info>');
+            $output->writeln('<info>*** NOTE: It is recommended to update Composer using "composer require ext-pdo" ***</info>');
             $output->writeln('<info>For more info, see: https://github.com/bayfrontmedia/bones/blob/master/docs/services/db.md</info>');
 
         }
@@ -487,9 +488,9 @@ class InstallService extends Command
 
                 ConsoleUtilities::msgInstalling($name, $output);
 
-                $src_file = Constants::get('BONES_RESOURCES_PATH') . '/cli/templates/install/service/veil/resources/views/examples/layouts/page.veil.php';
+                $src_file = Constants::get('BONES_RESOURCES_PATH') . '/cli/templates/install/service/veil/resources/views/examples/layouts/container.veil.php';
 
-                $dest_file = App::resourcesPath('/views/examples/layouts/page.veil.php');
+                $dest_file = App::resourcesPath('/views/examples/layouts/container.veil.php');
 
                 ConsoleUtilities::copyFile($src_file, $dest_file);